requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
pyjwt==2.9.0
argon2-cffi==23.1.0
